requirements.txt
# Core API
fastapi
uvicorn
python-dotenv
python-multipart

# LangChain / LangGraph
langchain
langchain-community
langchain-text-splitters
langchain-openai
langchain_huggingface
langchain-groq
langchain-google-genai
langchain-ollama
langgraph

# Vector Store
faiss-cpu  # fallback vector store, only needed if milvus_demo.db fails
pymilvus>=2.4.0
milvus-lite>=2.4.0

# ML / Feature Extraction
scikit-learn
pandas
numpy
joblib
xgboost>=2.0.0
lightgbm>=4.0.0
imbalanced-learn>=0.11.0

# NLP
transformers
torch
sentencepiece
sentence_transformers
rank_bm25
spacy

# Utilities
pymupdf
shap
matplotlib
plotly
reportlab
python-docx
deep-translator
